requirements.txt
Flask==3.0.0
opencv-python-headless==4.8.1.78
python-dotenv==1.0.0
Werkzeug==3.0.1
Pillow==10.1.0
pytest==7.4.3
gunicorn==21.2.0
